requirements.txt
Transformers              4.46.1
Pytorch                   2.5.1+cu124
torchvision               0.12.0
Datasets                  3.1.0
Tokenizers                0.20.1
accelerate                1.2.0
evaluate                  0.4.3
matplotlib                3.9.3
numpy                     2.1.3
opencv-python             4.10.0.84
pandas                    2.2.3
pillow                    11.0.0
pytorchvideo              0.1.5
safetensors               0.4.5
scikit-learn              1.6.0